update to latest changes

+++ trunner/device.py
@@ -1,3 +1,4 @@
+import importlib
 import logging
 import os
 import signal
@@ -10,10 +11,7 @@ import pexpect
 import pexpect.fdpexpect
 import serial
 
-try:
-    import RPi.GPIO
-except (ImportError, RuntimeError):
-    pass
+from pexpect.exceptions import TIMEOUT, EOF
 
 from .config import PHRTOS_PROJECT_DIR, DEVICE_SERIAL
 from .tools.color import Color
@@ -72,6 +70,14 @@ class Psu:
         if self.proc.returncode != 0:
             logging.error(f'Command {" ".join(self.proc.args)} with pid {self.proc.pid} failed!\n')
             raise Exception('Flashing IMXRT106x failed')
+
+
+def phd_error_msg(message, output):
+    msg = message
+    msg += Color.colorify('\nPHOENIXD OUTPUT:\n', Color.BOLD)
+    msg += output
+
+    return msg
 
 
 class PhoenixdError(Exception):
@@ -133,10 +139,10 @@ class Phoenixd:
 
         if self.wait_dispatcher:
             # Reader thread will notify us that message dispatcher has just started
-            dispatcher_ready = self.dispatcher_event.wait(timeout=10)
+            dispatcher_ready = self.dispatcher_event.wait(timeout=5)
             if not dispatcher_ready:
                 self.kill()
-                msg = f'message dispatcher did not start! Phoenixd output:\n{self.output()}'
+                msg = 'message dispatcher did not start!'
                 raise PhoenixdError(msg)
 
         return self.proc
@@ -149,9 +155,10 @@ class Phoenixd:
         return output
 
     def kill(self):
-        os.killpg(os.getpgid(self.proc.pid), signal.SIGTERM)
+        if self.proc.isalive():
+            os.killpg(os.getpgid(self.proc.pid), signal.SIGTERM)
         self.reader_thread.join(timeout=10)
-        if self.reader_thread.is_alive():
+        if self.proc.isalive():
             os.killpg(os.getpgid(self.proc.pid), signal.SIGKILL)
 
     def __enter__(self):
@@ -258,7 +265,7 @@ class Runner:
     """Common interface for test runners"""
 
     def flash(self):
-        """Method used for flashing device with image containing tests."""
+        """Method used for flashing a device with the image containing tests."""
         pass
 
     def run(self, test):
@@ -297,16 +304,17 @@ class GPIO:
 
     def __init__(self, pin):
         self.pin = pin
+        self.gpio = importlib.import_module('RPi.GPIO')
 
-        RPi.GPIO.setmode(RPi.GPIO.BCM)
-        RPi.GPIO.setwarnings(False)
-        RPi.GPIO.setup(self.pin, RPi.GPIO.OUT)
+        self.gpio.setmode(self.gpio.BCM)
+        self.gpio.setwarnings(False)
+        self.gpio.setup(self.pin, self.gpio.OUT)
 
     def high(self):
-        RPi.GPIO.output(self.pin, RPi.GPIO.HIGH)
+        self.gpio.output(self.pin, self.gpio.HIGH)
 
     def low(self):
-        RPi.GPIO.output(self.pin, RPi.GPIO.LOW)
+        self.gpio.output(self.pin, self.gpio.LOW)
 
 
 class IMXRT106xRunner(DeviceRunner):
@@ -350,6 +358,8 @@ class IMXRT106xRunner(DeviceRunner):
         try:
             with PloTalker(self.port) as plo:
                 plo.wait_prompt()
+                # FIXME We should wait for usb0 dev
+                time.sleep(1)
                 with Phoenixd(self.phoenixd_port) as phd:
                     plo.copy_file2mem(
                         src='usb0',
@@ -357,11 +367,11 @@ class IMXRT106xRunner(DeviceRunner):
                         dst='flash1',
                         off=0
                     )
-        except (pexpect.exceptions.TIMEOUT, pexpect.exceptions.EOF, PloError) as exc:
+        except (TIMEOUT, EOF, PloError, PhoenixdError) as exc:
             exception = f'{exc}\n'
             if phd:
-                exception += Color.colorify('\nPHOENIXD OUTPUT:\n', Color.BOLD)
-                exception += phd.output()
+                exception = phd_error_msg(exception, phd.output())
+
             logging.info(exception)
             sys.exit(1)
 
@@ -376,10 +386,15 @@ class IMXRT106xRunner(DeviceRunner):
             with PloTalker(self.port) as plo:
                 self.boot()
                 plo.wait_prompt()
+
+                if not test.exec_cmd:
+                    # We got plo prompt, we are ready for sending the "go!" command.
+                    return True
+
                 with Phoenixd(self.phoenixd_port, dir=load_dir) as phd:
-                    plo.app('usb0', test.exec_bin, 'ocram2', 'ocram2')
-        except (pexpect.exceptions.TIMEOUT, pexpect.exceptions.EOF, PloError) as exc:
-            if isinstance(exc, PloError):
+                    plo.app('usb0', test.exec_cmd[0], 'ocram2', 'ocram2')
+        except (TIMEOUT, EOF, PloError, PhoenixdError) as exc:
+            if isinstance(exc, PloError) or isinstance(exc, PhoenixdError):
                 test.exception = str(exc)
                 test.fail()
             else:  # TIMEOUT or EOF
@@ -387,8 +402,7 @@ class IMXRT106xRunner(DeviceRunner):
                 test.handle_pyexpect_error(plo.plo, exc)
 
             if phd:
-                test.exception += Color.colorify('\nPHOENIXD OUTPUT:\n', Color.BOLD)
-                test.exception += phd.output()
+                test.exception = phd_error_msg(test.exception, phd.output())
 
             return False
 
@@ -430,10 +444,14 @@ class HostRunner(Runner):
         if test.skipped():
             return
 
-        test_path = PHRTOS_PROJECT_DIR / f'_boot/{test.target}/{test.exec_bin}'
-
+        test_bin = PHRTOS_PROJECT_DIR / '_boot' / test.target / test.exec_cmd[0]
         try:
-            proc = pexpect.spawn(str(test_path), encoding='utf-8', timeout=test.timeout)
+            proc = pexpect.spawn(
+                str(test_bin),
+                args=test.exec_cmd[1:],
+                encoding='utf-8',
+                timeout=test.timeout
+            )
         except pexpect.exceptions.ExceptionPexpect:
             test.handle_exception()
             return
